Updated Create Datbase sql file with "thanks" Tables
two new tables were added to allow the possibility to
be able to thank users after their help. The idea is to add
a way for mentees who get help to thank the mentor. This new
thanks count will be featured on user profile to showcase some
sort of sense of achievement.

The reason for having two tables:
- "thanksCount" will keep count of thanks on individual user
  profile

- "whoThankedWho" will keep track of thanker and thankee and
avoid same user thanking twice